Made code much more modularized. Setting up for picking
The code's much more modularized now. For example, the polycube class doesn't depend on the scene to exist to work anymore.

Also setting up for polycube picking. Each polycube will have its own set of picking scenes built in. When we need to draw anyone of them, we simply switch the picking scene in the scene handler to the corresponding scene.

TODO:
-Test to see if this idea actually works.
-Set up what happens when a face is picked
-Set up polycube picking
-Set up hinge adjacency