Setting up hinge adjacency
It's been a real pain in the butt because I can't do simple mathematics.

In any case, hinge adjacency is ready to be set up. The caveat is because of how edges are stored, there are some special calculations that need to be made when adding adjacent edges into the list. Namely, hinges are not edges, but instead a pair of edges.

The easiest thing to do is calculate adjacency for the edges, ignoring the fact that they are not actually hinges, and then simply doing an 'incidence' check when necessary. To make this work efficiently, I'd need to make sure that hinges (in the context of the Hinge map), implicitly represent two edges. To do this, I think all I'd have to do is set an 'incident edge' parameter to a hinge object (easily done in javascript).