Found the Source of the Elusive Picking Bug
The toolbar pushed the canvas down so the picking texture was larger than the actual viewport. Woo.

Fixed by setting the toolbar to an absolute position with 0 padding at the top in CSS. Now the toolbar hangs over the canvas, which is a bit of a bother, but this makes the actual size of the texture independent of things such as padding, which is nice.