Fix to Face Adjacency
The bug reported just an hour before this message is fixed. The issue was with the way adjacency between diagonal cubes was checked. A cube would check diagonality (is this even a word) when it is placed, but not afterwards (duh). This means that the order in which a cube is placed in the polycube affected adjacency (which is bad).

The fix was simple: when a new cube is added, update the cubes around it to check for diagonality.

TODO:
-Fix bug for hinge adjacency. This bug is a bit more nuanced. It is not enough to check the edges between the adjacent faces as previously thought. If doing a second check for diagonality (see the fix that's being added in this patch), then there is a possbility that faces that are not adjacent are still adjacent to certain edges. It's a hard one to envision since the case is actually rather specific. The point is, a more thorough adjacency check must be made.