Fix to Face Adjacency
The bug reported just an hour before this message is fixed. The issue was with the way adjacency between diagonal cubes was checked. A cube would check diagonality (is this even a word) when it is placed, but not afterwards (duh). This means that the order in which a cube is placed in the polycube affected adjacency (which is bad).

The fix was simple: when a new cube is added, update the cubes around it to check for diagonality.