Fix to Edge Adjacency; Other changes
Here is a fix to edge adjacency. It's a slow one that checks each edge of two cubes against each other. It'll do for now, and it's a safe, sure-fire algorithm.

Also, there are some minor changes to the PolyCube class. There is now a Color2Cube map that'll work for picking cubes, and the AddCube function now allows for a custom id parameter, which will work for when we remove cubes.

In addition, the Toolbar_Handler class is set up with a right sidebar. It's not working properly for reasons unknown to me at the moment, but it is meant as a context menu for the different contexts a polycube may be in.