Fixed Hinge Line Bug
Fixed an issue where hinge lines were being malformed. By 'malformed' I mean sometimes lines didn't have cuts at their endpoints. This observation was rather innacurate because it was not that the lines didn't have cuts at that their endpoints. Instead, edges that were checked previous to finding a valid hinge line might be added to the hinge line because they were prematurely added to the line. The fix was just checking after recursion if a cut was found at any point in the line building (i.e. the other end of the line from the starting cut). If not, the line is simply scrapped.