Fixed back-end
The back-end of the polycube project (that is, all the stuff involving the polycube class and cube class) was a mess. The polycube class contained references to THREE.Scene() objects, holding some of it's picking scenes when it most certainly had no business doing that. The cube class also contained the THREE.Object3D classes that held the meshes for the cube parts, making accessing those meshes for rotation a complete hassle. Now things are modularized: the polycube class simply holds the data, the cube class just tells us what faces are missing in case we need them back, and a new visualizer class contains all the data pertaining to meshes