Cleaned Up Front-End; Fixed Scene Update Bugs
LESSON: This is probably a no duh, but object matrices are updated ON RENDER, which explains many of the bizarre bugs with objects not moving.

Fixed bugs regarding rotation polycubes not updating properly. Scenes were holding outdated versions of the polycubes, while the variables that I use to reference them has the new versions. Again, this seems like a no duh, but I guess it is an important note to keep.

The front end has been cleaned up a little. Updating cuts and hinges has now been relegated to being a function called in the update function every frame if the proper flags are set to true.

Rotations are halfway implemented. Need to add support for rotating faces that have already been rotated around hinges.

TODO:
-Add support for rotating faces
-Give user the ability to choose which direction they want to rotate