Rotations are close!
Using quaternions, we're able to use rotations. The problem was that rotations are not based on the world matrix, but rather the object's local matrix.

However, there are some bug fixes to take care of.

First, we need to be able to update edge axes as they are being rotated. If an edge that was aligned to the z-axis were to rotate along the y-axis, for example, it would end up aligned to the x-axis. Obviously, this is bad if we want to complex rotations.

Next, we need to find a way to allow the user to decide which way a rotation is done, as well as figure out a way to limit rotation so faces do not intersect.