Data loss bug fixed; taping is still buggy
Taping faces works in one sense, but it doesn't work in another sense. It works in the sense that I can now change what faces are neighbored to each other. What doesn't work is hinging. There is a bug where hinges are still being formed where hinges should not be. Not sure of the cause. Will need to be fixed ASAP