Optimizing Polycube class and helper classes
The app has tremendous slowdown when about 30 cubes are added. This could be due to a lack of optimization in the graphics, and also due to a lack of optimization in space usage.

First thing to do: fix up the polycube class and its helper classes. The cube class can be completely removed, and its duties can be given to the flexifaceedge graph. Having a map for locations to edge names can make things easier when it comes to finding adjacency between faces. The current algorithm for taping is O(n^2) where n is the amount of faces in the polycube. By simply mapping faces to their edges, mapping the edges to their locations, and then querying for other edges at that location, we can make that algorithm O(n).

Caching face graphs and doing some small optimization in creating cut trees in the dual graph is also a possible optimization.