Another fix to rotation bug
The issue was that the calculation to decide which direction to rotate faces depended on what hinge was clicked on. This is bad because the calculation is already invariant over what face was clicked on. It should've also been invariant on the hinge clicked on to ensure that the calculation is the same for a rotation line no matter what.