New module, some directory restructuring, streamlining
Added a new module that contains class definitions for edge nodes and face nodes in a dual graph.
Streamlined the process of adding faces. Checking for face incidence will now be a matter of checking if a face exists in a 3D array. While adding a new face to the dual graph, we need to give a position on the lattice. We then check if there is already a face at that position. If there is, the existing face is destroyed, and the new face will never be created.
The directory was restructured to be easier to navigate.