Added test manager, fixed object referencing issue, modified dual graph
Added a test manager class that can be used to test individual modules.
Fixed an issue where private functions would reference the wrong objects. For some reason, functions would keep old bindings, but not overwrite their new ones. There may be something I don't understand about JavaScript. I decided to have each private function that requires a 'this' reference to have the first parameter be the object to be referenced. Not the most graceful solution, but it will do for now.
The dual graph now holds face and edge data. Next is handling adjacency. After that, it should be relatively smooth sailing.
Deleted node modules folder. There is no need for it.